backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
motor==3.3.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
joblib==1.3.2
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0